Code Refactoring: Eliminate Duplication with Bearer Token Middleware
- Create authentication middleware (src/routes/auth-middleware.ts)
  * Handles Bearer token parsing and validation
  * Centralizes error handling for authentication failures
  * Supports both value-returning and void-returning service functions

- Refactor users-route.ts endpoints to use middleware
  * GET /api/users/current → bearerAuth(getCurrentUser)
  * DELETE /api/users/logout → bearerAuth(logoutUser)
  * Reduced 92 lines to 59 lines (-35%)
  * Eliminated 30+ lines of duplicated authentication code

- Benefits:
  * Single source of truth for Bearer token authentication
  * Consistent error handling across all auth endpoints
  * Easier to maintain and extend
  * Reduced bug risk from code duplication
