+++ src/routes/users-route.ts
@@ -1,5 +1,6 @@
 import { Elysia, t } from 'elysia';
 import { registerUser, loginUser, getCurrentUser, logoutUser } from '../service/users-service';
+import { bearerAuth } from './auth-middleware';
 import { db } from '../db';
 import { users } from '../db/schema';
 
@@ -54,39 +55,5 @@ export const usersRoute = new Elysia()
       })
     };
   })
-  .get('/api/users/current', async ({ set, headers }: any) => {
-    const authHeader = headers['authorization'] || headers['Authorization'];
-
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-      set.status = 401;
-      return { error: 'Unauthorized' };
-    }
-
-    const token = authHeader.substring(7);
-
-    try {
-      const user = await getCurrentUser(token);
-      return { data: user };
-    } catch (err: any) {
-      set.status = 401;
-      return { error: 'Unauthorized' };
-    }
-  })
-  .delete('/api/users/logout', async ({ set, headers }: any) => {
-    const authHeader = headers['authorization'] || headers['Authorization'];
-
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-      set.status = 401;
-      return { error: 'Unauthorized' };
-    }
-
-    const token = authHeader.substring(7);
-
-    try {
-      await logoutUser(token);
-      return { data: 'OK' };
-    } catch (err: any) {
-      set.status = 401;
-      return { error: 'Unauthorized' };
-    }
-  });
+  .get('/api/users/current', bearerAuth(getCurrentUser))
+  .delete('/api/users/logout', bearerAuth(logoutUser));
