--- test-logout.ts
@@ -1,1 +1,0 @@
-﻿console.log(" Test file created\);
